[refactor] odemis.gui.acqmng.py has been renamed to preset.py file
The python file contains functions related to preset which are used during acquisition.
